Added the ability to find keyshares by iterating inside every function
This was not implemented in a serious way, but just to see if it can
work. A pointer iterates over every function, from its start point to
its RET point. It detects the RET because I insert special keyshare
values before (this can change later). SOS: What if we have many RETs in
one function?