Memory manager: Added Actual high-level manager with free() support
Added high level memory manager, uses linked lists to keep track of
memory. Suffers from external fragmentation (a fix can be to compact
memory whenever asked, but that would not allow the user to do low level
operations in memory, as compact() would change the location of the
user's data - even though it would not change the data).