Secure stack manager: Bug fixes, added the ability to pass byte blobs:
Made sure to set every pointer to NULL if uninitialized (free() goes
crazy otherwise). 
Also, added the ability to pass to function arguments (or declare
locally in a function) an arbitrary blob of bytes, which can then be
interpreted in any way that the user pleases. This functionality may
render unnecessary the obligation to include other ways of passing
arguments (like array arguments) later. The user can use the blob of
bytes to pass anything.