Added more user-friendliness.
Added user friendliness to secure stack manager. Also, pushed the duty
of freeing the arb_pointer parameters (that are stored in the normal
heap) to the user. The stack manager will not try to free these (as they
may be global variables, and in that case we would have undefined
behavior).